+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ## [4.5.2] - 2025-05-06
 
 ### Fixed
-- mongo_class.Server.unlock_db: Changed self.conn.unlock to calling adm_cmd method to unlock the database.
+- mongo_class.Server.lock_db: Replaced fsync() with admin.command() call.
 - Renamed mongo.py to mongo_template.py due to conflict in testing.
 
 ## Changes